refactor main screen scrolling and editor items for lazy scope
